pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "project-sms-send"
version = "0.0.2"
description = ""
authors = ["Anton Perepelitcyn <antonio.puh@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.0"
phonenumbers = "^8.13.36"
celery = "^5.4.0"
gevent = "^24.2.1"
pytz = "^2024.1"
redis = "^5.0.4"
flake8-pyproject = "^1.2.3"
django-stubs = "^5.0.2"
django-stubs-ext = "^5.0.2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^4.4"
mypy = "^1.10.0"
pytest = "^7.2.2"
pytest-django = "^4.8.0"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
isort = "^5.13.2"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"

[tool.isort]
profile = 'django'
quiet = true
include_trailing_comma = true
use_parentheses = true
src_paths=['backend']
multi_line_output = 5
known_django = 'django'
sections = 'FUTURE, STDLIB, DJANGO, THIRDPARTY, FIRSTPARTY, LOCALFOLDER'
default_section = 'THIRDPARTY'
line_length = 80
skip_glob="*/migrations/*"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'app.settings'
env = [
  'SECRET_KEY=your-secret-key',
  'CACHE_URL=dummycache://',
]
markers = [
  "override_settings: does the same as `django.test.override_settings` but for pytest.",
]

[tool.flake8]
show-source = true
statistics = false
doctests = true
enable-extensions = 'G'
max-line-length = 100
max-complexity = 6
max-imports = 15
ignore = [
  'W605',
  'W504',
]
exclude = [
  '.git',
  '__pycache__',
  '.venv',
  '.env',
  './*/migrations/*',
  'tmp',
  'util',
  'logs',
  'media',
  # no need to check settings
  'app/settings.py',
]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
plugins = [
  'mypy_django_plugin.main',
]
exclude = [
  '/?migrations/.*\.py$',
  '/?tests/.*/?test\_.*\.py$',
]

[tool.django-stubs]
django_settings_module = 'app.settings'
strict_settings = false
